search / test swagger

+++ app/Http/Controllers/EpisodeController.php
@@ -7,6 +7,9 @@ use App\Http\Requests\UpdateEpisodeRequest;
 use App\Models\Episode;
 use App\Models\Podcast;
 use CloudinaryLabs\CloudinaryLaravel\Facades\Cloudinary;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Http\Request;
+
 
 class EpisodeController extends Controller
 {
@@ -15,6 +18,7 @@ class EpisodeController extends Controller
      */
     public function index(Podcast $podcast)
     {
+
         try {
             $episodes = Episode::where('podcast_id', $podcast->id)
                 ->with('podcast')
@@ -34,19 +38,20 @@ class EpisodeController extends Controller
      */
     public function store(StoreEpisodeRequest $request, Podcast $podcast)
     {
+        Gate::authorize('is-owner', $podcast);
+
         try {
             $infos = $request->validated();
 
-            if ($request->hasFile('audio')) {
 
-                $uploadedFile = Cloudinary::uploadFile(
-                    $request->file('audio')->getRealPath(),
-                    ['resource_type' => 'auto']
+            if ($request->hasFile('audio_file')) {
+                $uploadedAudio = Cloudinary::upload(
+                    $request->file('audio_file')->getRealPath(),
+                    ['resource_type' => 'auto' ]
                 );
 
-                $infos['audio'] = $uploadedFile->getSecurePath();
+                $infos['audio_file'] = $uploadedAudio->getSecurePath();
             }
-
             $episode = $podcast->episodes()->create($infos);
 
             return [
@@ -60,7 +65,6 @@ class EpisodeController extends Controller
         }
     }
 
-
     /**
      * Display the specified resource.
      */
@@ -69,24 +73,23 @@ class EpisodeController extends Controller
         return $episode;
     }
 
-
     /**
      * Update the specified resource in storage.
      */
     public function update(UpdateEpisodeRequest $request, Episode $episode)
     {
+        Gate::authorize('is-owner', $episode);
         try {
 
-             $infos = $request->validated();
+            $infos = $request->validated();
 
-            if ($request->hasFile('audio')) {
-
-                $uploadedFile = Cloudinary::uploadFile(
-                    $request->file('audio')->getRealPath(),
+            if ($request->hasFile('audio_file')) {
+                $uploadedAudio = Cloudinary::upload(
+                    $request->file('audio_file')->getRealPath(),
                     ['resource_type' => 'auto']
                 );
 
-                $infos['audio'] = $uploadedFile->getSecurePath();
+                $infos['audio_file'] = $uploadedAudio->getSecurePath();
             }
 
             $episode->update($infos);
@@ -107,16 +110,29 @@ class EpisodeController extends Controller
      */
     public function destroy(Episode $episode)
     {
+        Gate::authorize('is-owner', $episode);
         try {
             $episode->delete();
 
             return [
                 'message' => 'Episode supprimé avec succès',
             ];
-        } catch (\Exception $th) {
+        } catch (\Exception $e) {
             return [
-                'error' => $th->getMessage()
+                'error' => $e->getMessage()
             ];
         }
     }
+
+        public function search(Request $request)
+{
+    $search = $request->input('query');
+
+    $podcasts = Episode::where('title', 'LIKE', "%$search%")
+        ->orWhere('description', 'LIKE', "%$search%")
+        ->orWhere('duree', 'LIKE', "%$search%")
+        ->get();
+
+    return response()->json($podcasts);
+}
 }
